requirements.txt
psycopg2-binary==2.8.1
django-background-tasks==1.2.0
celery==4.3.0
djangorestframework==3.9.2
openpyxl==2.4.9
openpyxl-templates==0.1.24
